FLUX: support huggingface dataset
